requirements.txt
hikari>=2.0.0.dev106
hikari-lightbulb>=2.2.0
python-dotenv>=0.19.2
lavasnek-rs>=0.1.0a4
uvloop>=0.16; sys_platform != "win32"